App: Changed LSS Cold startup behaviour...
we now wait until we are LSS waiting before starting CANopenNode init.
This fixes problems where LSS switch waiting message was lost because
CANopenNode Init function was manipulating filter settings in CAN
hardware. This caused two nodes getting the same node ID assinged as
this node was still selected when the next node in the line was assigned
an ID.